reorg splunk files
